feat: Prevent multiple app instances
Adds functionality to prevent multiple instances of the application from running simultaneously.

This is achieved by using a local server. Upon startup, the application checks if a server with a specific name already exists. If it does, it means another instance is running, and the application displays an error message and exits. If no server exists, the application creates one, preventing subsequent instances from starting. This prevents data corruption and resource conflicts that can occur when multiple instances operate on the same data.